requirements.txt
Package                      Version
---------------------------- ---------
absl-py                      1.1.0
asttokens                    2.0.5
astunparse                   1.6.3
autokeras                    1.0.19
backcall                     0.2.0
cachetools                   5.2.0
certifi                      2022.6.15
charset-normalizer           2.0.12
click                        8.1.3
colorama                     0.4.5
cv                           1.0.0
cycler                       0.11.0
debugpy                      1.6.0
decorator                    5.1.1
entrypoints                  0.4
executing                    0.8.3
Flask                        2.1.2
flatbuffers                  1.12
fonttools                    4.33.3
gast                         0.4.0
google-auth                  2.8.0
google-auth-oauthlib         0.4.6
google-pasta                 0.2.0
grpcio                       1.46.3
h5py                         3.7.0
idna                         3.3
importlib-metadata           4.11.4
ipykernel                    6.15.0
ipython                      8.4.0
itsdangerous                 2.1.2
jedi                         0.18.1
Jinja2                       3.1.2
joblib                       1.1.0
jupyter-client               7.3.4
jupyter-core                 4.10.0
keras                        2.9.0
Keras-Preprocessing          1.1.2
keras-tuner                  1.1.2
kiwisolver                   1.4.3
kt-legacy                    1.0.4
libclang                     14.0.1
llvmlite                     0.38.1
Markdown                     3.3.7
MarkupSafe                   2.1.1
matplotlib                   3.5.2
matplotlib-inline            0.1.3
nest-asyncio                 1.5.5
numba                        0.55.2
numpy                        1.22.4
oauthlib                     3.2.0
opencv-python                4.6.0.66
opt-einsum                   3.3.0
packaging                    21.3
pandas                       1.4.2
parso                        0.8.3
pickleshare                  0.7.5
pip                          22.1.2
prompt-toolkit               3.0.29
protobuf                     3.19.4
psutil                       5.9.1
pure-eval                    0.2.2
pyasn1                       0.4.8
pyasn1-modules               0.2.8
Pygments                     2.12.0
pyparsing                    3.0.9
python-dateutil              2.8.2
pytz                         2022.1
pywin32                      304
pyzmq                        23.2.0
requests                     2.28.0
requests-oauthlib            1.3.1
rsa                          4.8
scikit-learn                 1.1.1
scipy                        1.8.1
seaborn                      0.11.2
setuptools                   58.1.0
six                          1.16.0
sklearn                      0.0
stack-data                   0.3.0
tensorboard                  2.9.1
tensorboard-data-server      0.6.1
tensorboard-plugin-wit       1.8.1
tensorflow                   2.9.1
tensorflow-estimator         2.9.0
tensorflow-io-gcs-filesystem 0.26.0
termcolor                    1.1.0
threadpoolctl                3.1.0
tornado                      6.1
traitlets                    5.3.0
typing_extensions            4.2.0
urllib3                      1.26.9
wcwidth                      0.2.5
Werkzeug                     2.1.2
wheel                        0.37.1
wrapt                        1.14.1
zipp                         3.8.0
Pillow                       9.1.1
